requirements.txt
# AI Refactoring Agent API Dependencies;

Web Server;

fastapi;
uvicorn;

AI SDK;

google-genai;